CYBERSÉCURITÉ/TP3/Fixes/XSS/requirement.txt
Flask==3.0.0
requests
pyjwt
bleach
sendgrid